lib: Implement snowflakes
This commit adds the initial snowflake implementation. Specifically, it
adds a `Snowflake` type that represents individual snowflake IDs,
a `SnowflakeGenerator` type to generate new snowflakes, and a
`SnowflakeComparator` type that allows comparing snowflakes with
timestamps. Moreover, this commit adds loads of documentation to fully
describe the existing API and help users of the crate find their way
around its features.

This commit also adds a rustfmt configuration file that overrides the
maximum line width to be more appropriate for modern displays.

Finally, this commit adds a Cargo configuration to allow building the
crate and including it in other crates as a dependency.