lib: Relax lock-free generator implementation
This commit makes the initial load operation of the lock-free generator
implementation `Relaxed`. We can do this, as the CAS operation already
establishes an inter-thread happens-before relationship by synchronizing
with other threads running the lock-free generator implementation.

In addition to changing the memory ordering for the initial load
operation, this commit adds comments explaining the other ordering used
to synchronize with other threads in the CAS operation.