benches: Implement generator benchmark
This commit adds a benchmark to test generators. As the current
implementations' performance exceeds the limits of the provided
`ClassicLayout` (its sequence number is too small), this benchmark uses
a simplified layout that consists of 42 bits for the timestamp and 22
bits for the sequence number.

The benchmarks use 10 threads to simulate contention problems.

The current blocking implementation can generate up to 6.2 million
snowflakes per second, and the lock-free implementation can generate up
to 16.2 million snowflakes per second (on modern hardware).