lib: Implement generator unit tests
This commit adds unit tests for the current generator implementations.
At least for the generators, this should reach 100% line coverage and
all important branches. Parts that we don't want to cover (e.g. because
they disappear at compile time) are marked with `// Skip coverage`
comments that explain why those parts are skipped.

To test non-monotonic clock behaviour and make the unit tests
independent of the current time, this commit adds a mocked system time.
This system time (alongside other utility functions that allow
manipulating a generator's state) are only compiled in unit tests, so
they don't change the public API or the compiled code if people use
Snowdon as a library.