lib: Implement snowflake unit tests
This commit implements unit tests for `Snowflake`'s various methods and
associated functions. Moreover, it tests the `(Partial)Ord` and
`(Partial)Eq` implementations to ensure snowflakes behave correctly when
compared with other snowflakes.

While we test a snowflake's comparator method, we don't try to verify
`Ord` or other trait implementations for `SnowflakeComparator` with the
tests added by this commit.