classic: Panic on invalid machine ID
This commit changes the `construct_snowflake` implementation of the
provided `ClassicLayout` to panic on invalid machine IDs. Previously, it
would generate invalid snowflakes if the machine ID exceeds the number
of bits available in the layout. With this commit, it should be easier
for developers to notice invalid machine ID implementations.