lib: Utilize modification order in lock-free gen
This commit makes the lock-free generator utilize
`last_snowflake_atomic`'s modification order to further relax the memory
ordering used in the compare and swap operation at the bottom of our
CAS-loop. This should improve performance of our atomic operations on
some platforms, as this avoids having to synchronize with other
generating threads. The algorithm implementation should remain correct,
however.

In a future commit, we'll try to verify the correctness of this change
with `loom`.