cargo: Bump criterion to 0.5
While this update doesn't change our current benchmarks, it increases
the MSRV for tests and benchmarks. This doesn't change our "production"
MSRV, however, as this dependency is only used as a dev-dependency.

However, this commit also adds a note to README.md explaining that users
might need a higher rust version when developing snowdon itself.