comparator: Replace get_comparator with TryFrom
This commit replaces `Snowflake::get_comparator` with a `TryFrom` (and
`TryInto`) implementation for `SnowflakeComparator`. This provides the
same functionality while matching Rust's API guidelines more closely.

Specifically, the previous method wasn't a "real" getter in the first
place. Instead, it (cheaply) converted a snowflake into a comparator.
Instead of simply renaming the method, I moved it to a trait
implementation entirely, as `as_comparator` would indicate that the
conversion is infallible.

This patch also updates the documentation and tests to use the new
implementation instead.