lib: Rename Snowflake getters
This commit changes the getters throughout the crate to match Rust's
preferred naming (without the `get_` prefix). Moreover, it changes the
`get` method in `Snowflake` (and the `get_i64` method in the classic
layout extension) to `into_inner`. Note that the parameter change from
`&Snowflake` to `Snowflake` doesn't impose any restrictions on using
code, as Snowflake implements `Copy`.