cargo: Remove explicit `serde` feature
This commit removes the explicit declaration of the `serde` feature.
Without any additional features enabled by it, there's no need to
explicitly declare the feature, as Cargo already creates a feature for
the optional dependency.

This patch also adds a comment stating that the `serde` feature is
automatically created for the optional dependency, as some users might
look at the `Cargo.toml` to determine whether a crate supports
(optional) Serde support.