lib: Implement (Try)From
This commit implements the `From` and `TryFrom` traits for `Snowflake`
and `SnowflakeComparator` for type conversions that make sense.
Notably, we don't implement it for
`SnowflakeGenerator::from_raw_timestamp`, as that would mislead users
that expect the function to take a snowflake integer representation
instead.

This commit doesn't remove the existing associated functions, however,
as those tend to be more readable than `into()` or `try_into()`, so
users might prefer them over the trait implementations.

Moreover, this patch adds assert statements to the existing unit tests
that verify that the `From` and `TryFrom` implementations match the
associated functions.