meta: Re-license Snowdon under MIT & Apache-2.0
This commit changes Snowdon's licence to a dual-licensing model of
Apache-2.0 (the old licence) and MIT (Expat License). The user is free
to choose one of these licences.

This change allows certain types of projects to include Snowdon as a
dependency more easily, as the MIT Expat License imposes fewer
restrictions.