meta: Add CHANGELOG.md
Now that we've released the first version of Snowdon, it's best to
maintain a changelog. This commit adds a changelog containing the
initial release notes for 0.1.0 and the current changes we'll release
as `0.2.0`.

For the breaking section, this changelog contains instructions on how
to update code that's affected by the change.