lib: Serialize snowflakes as u64
This commit makes the Serde-support implementation of `Deserialize` and
`Serialize` (de-)serialize snowflake IDs as u64 instead of a full
mapping (which even included the `_marker` field).

While being a breaking change, this is probably the behaviour a user
expects when serializing a numeric ID. Moreover, this matches the
`Display` implementation.